requirements.txt
fastapi
uvicorn[standard]
openai
pinecone-client
python-dotenv
pydantic
numpy
